CHROMIUM: pmdomain: mediatek: add checking for domain on
Some power domains are controlled by voting, but the completion of
the voting process does not guarantee that the power domain has been
powered on.
There are situations where the power domain is not actually powered on
after the return from power on call.
To fix the issue, we add a step to check the power domain is actually
powered on before power on call returns.

BUG=b:394161210
BUG=b:389800877
BUG=b:394326329
TEST=Boot up OK & Suspend/Resume OK & Create New video OK using Kinemaster
UPSTREAM-TASK=b:379034565

Signed-off-by: Guangjie Song <guangjie.song@mediatek.corp-partner.google.com>
Change-Id: I0799455dc0b02c95bc94dcfe87fb07e018858c2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6249278
Reviewed-by: Fei Shao <fshao@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>