CHROMIUM: ipu6: add Bus info
The bus info is used in tast.camera.KernelSmokeMIPI test. Add it back.

Since the media graph of the upstream ISYS driver is different from the
downstream one, we also need to update the test config and modificiation
to v4l2_media_test. The latter is needed to properly distinguish between
MIPI and USB camera.

UPSTREAM-TASK=b:149068439
UPSTREAM-TASK=b:149068672

BUG=b:394991505
TEST=tast run camera.KernelSmokeMIPI w/ new config \
	and modified v4l2_media_test

Change-Id: I519a8ad921fbc27c6c9403d05931acd8e5edf376
Signed-off-by: Hidenori Kobayashi <hidenorik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6246076
Reviewed-by: Sergey Senozhatsky <senozhatsky@chromium.org>