CHROMIUM: media: jpeg: Always reset jpgdec hw after decode done
The hardware does not consider the problematic input buffer
to be an error and continues to decode, which will cause
output loss. When the decoding is completed, the hardware
state needs to be reset immediately in the irq handler,
instead of waiting for the next input to be sent and then
doing it, otherwise the error of the problematic input buffer
cannot be handled.

BUG=b:396048398
TEST=build pass test jpegdec pass
UPSTREAM-TASK=b:379040143

Signed-off-by: Kyrie Wu <kyrie.wu@mediatek.corp-partner.google.com>
Change-Id: Ib992a1532f210a154eb28bb13834accbafd97b0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6267719
Reviewed-by: Fei Shao <fshao@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>