UPSTREAM: efi/cper: Fix cper_ia_proc_ctx alignment
According to the UEFI Common Platform Error Record appendix, the
IA32/X64 Processor Context Information Structure is a variable length
structure, but "is padded with zeros if the size is not a multiple
of 16 bytes".

Currently this isn't honoured, causing all but the first structure to
be garbage when printed. Thus align the size to be a multiple of 16.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
(cherry picked from commit 174e5e9da4f5946de3d09c32ee56cbbc9d70505b)

BUG=b:242492345
TEST=Guybrush

Change-Id: Ibd449eafbce29283c3f9dd38f329a3331fcc55cd
Signed-off-by: Zheng Bao <zheng.bao@amd.corp-partner.google.com>
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6914180
Reviewed-by: Lina Versace <linyaa@google.com>
Reviewed-by: Tim Van Patten <timvp@google.com>