CHROMIUM: iwl7000: pcie: Fix TSO preparation
The allocation of the scatter gather data structure should be done
based on the number of memory chunks that need to be mapped, and it
is not dependent on the overall payload length. Fix it.

In addition, as the skb_to_sgvec() function returns an 'int' do not
assign it to an 'unsigned int' as otherwise the error check would be
useless.

Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Change-Id: I3cb4d6768c9d28ce3da6cd0a6c65466176cfc1ee
(cherry picked from commit 5714ccffb33a783f0ccbf1516fb23064016ddf3c)
iwl7000-tree: 83459d4d6ea1870788b6df1a4f5d4ee1e2bca4c6