CHROMIUM: drm/bridge: anx7625: Move HDCP update back to .atomic_check()
Do a partial revert of commit 0e063a7e8262 ("FROMGIT: drm/bridge:
anx7625: Update HDCP content status") to move the HDCP-related works
back to .atomic_check() callback.

This was originally rejected bt the upstream reviewer, but this is the
only approach that works with the current Chrome implementation.

BUG=b:380167784
TEST=emerge-skywalker chromeos-kernel-6_6
UPSTREAM-TASK=b:402659708

Change-Id: Ib77b43f76873ce58a33e2e8847122f4e4b262b64
Signed-off-by: Pin-yen Lin <treapking@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6326054
Reviewed-by: Fei Shao <fshao@chromium.org>