CHROMIUM: Revert "drm/amd/display: Fix DSC-re-computing"
This reverts commit 3f9f631f9b910c4aeafbeaee6ef08a3c193f0c29.

The following commit in the v6.6 stable branch:

commit 3f9f631f9b910 ("drm/amd/display: Fix DSC-re-computing")

causes a kernel crash when an external hub with 2 monitors is attached.
This causes a crash loop for as long as the device has the hub attached.

These 2 patches touch the same code and revert cleanly to get us back to
a good state:

commit e8b8c1ecbd2cc ("drm/amd/display: Fix incorrect DSC recompute trigger")
commit 3f9f631f9b910 ("drm/amd/display: Fix DSC-re-computing")

BUG=b:400356211
TEST=guybrush/dewatt attached to an external doc with 2 monitors
UPSTREAM-TASK=b:403279793

Change-Id: Ie49dfbf445410bded8b61f4983e00d9fd6203923
Signed-off-by: Ross Zwisler <zwisler@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6353480
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Commit-Queue: Drew Davenport <ddavenport@chromium.org>