CHROMIUM: clocksource: Add parameter to allow some watchdog's skew tolerance
Introduce the parameter "cs_watchdog_tolerate_skew=N", where:
- if N > 0, configures the clocksource watchdog to allow N skew
  detections before effectively marking the clocksource as unstable;

- if N < 0, configures the clocksource watchdog to NEVER mark the
  clocksource as unstable, no matter how many skews are detected;

If the parameter is not passed (or N == 0) the watchdog behaves
as the default, marking the cs as unstable in the first detected
skew.

Also notice that while N > 0 (or if N < 0) we *always* show the
print on dmesg in case the watchdog read routine requires any
retries to have a proper watchdog clock read.

UPSTREAM-TASK=b:404286907
BUG=b:399422076
TEST=The easier way of testing this is to change the uncertainty
check on clocksource_watchdog() (kernel/time/clocksouce.c),
like the following patch:

-               if (abs(cs_nsec - wd_nsec) > md) {
+               if (abs(cs_nsec - wd_nsec) > 500) {

This will trigger the watchdog for sure and this way, this patch
could be tested. Creating a sysctl/parameter to tune the md value
is another good testing approach. Just remember of setting the
cs_watchdog_tolerate_skew=N in the command-line as well!

Reviewed-by: Thadeu Lima de Souza Cascardo <cascardo@igalia.com>
Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>
Co-developed-by: Helen Koike <koike@igalia.com>
Signed-off-by: Helen Koike <koike@igalia.com>
Change-Id: Idb3a61596e125be332894d01a439611690cdcb50
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6361307
Reviewed-by: Deepa Dinamani <deepadinamani@google.com>
Commit-Queue: Deepa Dinamani <deepadinamani@google.com>
Reviewed-by: Matt Davis <mattedavis@google.com>
Tested-by: Deepa Dinamani <deepadinamani@google.com>