FIXUP: CHROMIUM: pinctrl: mediatek: Update eint design (v2)
This essentially reverts CL:6342175 and reapply its PS1.

I misunderstood the code structure and thought it would have been better
to have a static get function for the entire mtk-eint driver, but it
turns out it's not needed, and even worse is that applying the helper
function implies we actually need signed integer for the return value.

For simplicity, deconstruct and expand the newly added helper function
to fix boot issue.

BUG=b:402630846,b:403768919
UPSTREAM-TASK=b:379036144
TEST=Boot Navi

Change-Id: Ia7eb5c1bb9b9bbcb9865fc3c7e71587b85d46aef
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6356750
Reviewed-by: Hsin-Te Yuan <yuanhsinte@chromium.org>