CHROMIUM: usb: typec: ucsi: Notify power_supply class of non-PD charger
When handling a power opmode change, the UCSI driver will only check
PDOs if the power opmode is PD. The functions to check PDOs the notify
the power supply class of a change. This updates ucsi_pwr_opmode_change
to also notify the power_supply class of a port change when registering
a non-PD charger.

BUG=b:399055364
UPSTREAM-TASK=b:353605100
TEST=Verified the kernel correctly alerted userspace of a non-PD charger
on Hylia.

Change-Id: Idc1453885f31e45b7a00c45f0fe1ad89e3c894ca
Signed-off-by: Jameson Thies <jthies@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6420171
Reviewed-by: Eric Yilun Lin <yllin@google.com>
Reviewed-by: Benson Leung <bleung@google.com>