CHROMIUM: drm/mediatek: Refine HPD flow to fix quick HPD change issue
To ensure that the DRM CRTC is closed when HPD is low.
It will cause the DRM not to open the DP if the HPD changes quickly,
because the DRM thinks that the DP is connected.

Besides, to ensure that the DP returns the correct connection status to
the DRM, add a check for mtk_dp->dp_ready in mtk_dp_con_detect_v2.

BUG=b:380379750,b:404648100,b:378606744
TEST=build ok/verify pass locally
UPSTREAM-TASK=b:379039307

Change-Id: I397c60ebd1c6f88da77ebac24aae12ac53ccfb5a
Signed-off-by: Ives Chenjh <ives.chenjh@mediatek.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6449237
Tested-by: Wentao Qin <qinwentao@huaqin.corp-partner.google.com>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>