CHROMIUM: img-rogue: Improved VA re-mapping policy
GPU VAs of buffers (Freelist, MLIST) could be remapped if it was
provided as a parameter to BridgeDevmemIntReserveRangeAndMapPMR.

In the improved implementatin, in DevmemIntMapPMR() a check will
be added to check if reservation is made on PMMETA_PROTECT heap.
If yes then prevent MMU_MapPMRFast code from re-mapping existing
valid entries.

IMG ID: PSP-60
BUG: b/394890268
TEST=graphics.DRM.*, graphics.WebGLAquarium.1000_fishes,
     graphics.WebGLManyPlanetsDeep, graphics.GLBench

Change-Id: I8c0c99733d80a7a2d43171375dda852837da8cbe
Signed-off-by: Faijur Rahman <faijur.rahman@imagination.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6470303
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Pin-yen Lin <treapking@chromium.org>
Tested-by: Pin-yen Lin <treapking@chromium.org>
Reviewed-by: Prahlad Kilambi <prahladk@google.com>