CHROMIUM: ipu6: psys: ignore the imported dmabufs vmap
For the imported dma buffers, psys driver is not expected to vmap
them to get the kernel virtual address as not all exported buffers
support that and the cache-coherency should be handled by the owner
of dma buffers instead of the importers.

This patch ignore the dma buffer vmap for imported dma buffers.

UPSTREAM-TASK=b:149068439
UPSTREAM-TASK=b:149068672

BUG=b:430172734
BUG=b:417882826
TEST=camera smoke test in nissa

Fixes: 10f233ebab6c ("CHROMIUM: ipu6: call ipu6 DMA mapping in psys instead of dma_ops")
Change-Id: I523b8026561dacb1c213b7822e7012e4554961dd
Signed-off-by: Bingbu Cao <bingbu.cao@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6654553
Tested-by: Yudhistira Erlandinata <yerlandinata@chromium.org>
Reviewed-by: Ricardo Ribalda <ribalda@chromium.org>
Reviewed-by: Yudhistira Erlandinata <yerlandinata@chromium.org>
Commit-Queue: Ricardo Ribalda <ribalda@chromium.org>