CHROMIUM: iwl7000: Don't read SMEM when collecting debug data
This is a large amount of data, which slows down the reset and results in ping loss. To workaround this, don't read the SMEM region. Since this only happens with an HR and GF, limit this workaround to it.

Signed-off-by: Arowa Suliman <arowa@google.com>

BUG=b:410088183
TEST=Ran wifi.Reset.* tests on Brask

Cq-Depend: chromium:6496006
Cq-Depend: chromium:6494954
Cq-Depend: chromium:6496007
Cq-Depend: chromium:6490054
Cq-Depend: chromium:6496008
Change-Id: I715fa76696a7a8b3ba516d2f55a4399bf34bc396
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6494953
Commit-Queue: Arowa Suliman <arowa@chromium.org>
Tested-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Ruth Mekonnen <rmekonnen@google.com>