CHROMIUM: remoteproc: apu: set apu hw log water level
The NPU will generate an interrupt to the APMCU if the log buffer
reaches the set water level. If the value is not set, the interrupt
will be generated when the buffer is full.

The size of the log buffer is 1MB, thus we simply set the water level
to half.

UPSTREAM-TASK=b:379040514
BUG=b:405671135
TEST=emerge-rauru chromeos-kernel-6_6

Signed-off-by: Chungying Lu <chungying.lu@mediatek.corp-partner.google.com>
Change-Id: I2b59777134925c6a12201bd556287316878d223d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6491214
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>