CHROMIUM: Revert "HID: Ignore battery for all ELAN I2C-HID devices"
This reverts commit 6ec7c739346ce9eeaa02fc2bb1d5119acfa1102a.

The following commit in the v6.6 stable branch:

commit 6ec7c739346ce ("HID: Ignore battery for all ELAN I2C-HID devices")

disables stylus battery monitoring for all Elan touchscreens. This
prevents ChromeOS from producing low stylus battery notifications
or show battery level indicators on any Chromebook that uses an Elan
touchscreen.

The change was motivated by the general Linux laptop experience
outside the ChromeOS ecosystem, and is not appropriate for our use.
The behaviour prior to the change is fine, as the single Elan exceptions
are for firmware not used by Chromebooks.

BUG=b:405191173
TEST=redrix/brya w/ ELAN tested with USI stylus
UPSTREAM-TASK=b:416313411

Change-Id: I27c0deff7a4c6d25f3242da6a0ae2d98615f8a03
Signed-off-by: Kenneth Albanowski <kenalba@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6521040
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>