CHROMIUM: remoteproc: apu: Add power reference count protection
We use a power reference count to control when to power on or off the
NPU. Additionally, we use a timer to delay power off for power saving,
avoiding frequent power on/off cycles when the NPU is performing
successive jobs.

However, there might be some operations that call NPU power on/off,
such as the devfreq framework. These operations will call the
power-on/off API directly when their operation is done to power off the
NPU.

The reference count will be referenced and updated in both the
power-on/off API and the delay power off timer. Therefore, a mutex
protection is needed to avoid race conditions in the power reference
count.

UPSTREAM-TASK=b:379040514
BUG=b:402755110
TEST=emerge-rauru chromeos-kernel-6_6

Signed-off-by: Chungying Lu <chungying.lu@mediatek.corp-partner.google.com>
Change-Id: I5761786476237c5b80a34784a8a9d6a1379b17a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6512115
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>