FROMGIT: drm/i915/dsi: Enforce pipeline flush with DSI HS transfer
With all of the boundary conditions when streaming the commands B2B in our
validation (part of the reason we added the flush),  the Flush effectively
serializes the transmission of each command enqueued within the command
dispatcher to one per V. Blank line which simplifies the behavior of the
High Speed Arbitration.

So, unless we absolutely have to burst these to the Sink, we should be
using the Pipeline Flush bit to serialize the commands.

Bspec: 19742, 50193
Closes: https://gitlab.freedesktop.org/drm/i915/kernel/-/issues/14247
Cc: Suraj Kandpal <suraj.kandpal@intel.com>
Cc: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Gareth Yu <gareth.yu@intel.com>
Link: https://lore.kernel.org/r/20250509092539.763389-1-gareth.yu@intel.com
Signed-off-by: Jani Nikula <jani.nikula@intel.com>
(cherry picked from commit 1c57014325ef2b62459c8482768a842f9c40cd0c
 https://gitlab.freedesktop.org/drm/i915/kernel.git drm-intel-next)

BUG=b:411317352
TEST=Change DSI backlight with brightness up and down key quickly

Change-Id: I51cb18d189a7f783e06a04ef5e797b6c76ac373b
Signed-off-by: Gareth Yu <gareth.yu@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6538914
Reviewed-by: Edward Doan <edoan@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Ricky Chang <rickytlchang@chromium.org>