CHROMIUM: fanotify: don't leak group ref-count on error
We forget to roll-back fsnotify_prepare_user_wait()
changes when FANOTIFY_PERM_EVENTS preparation races
with mark creation/destruction, which makes it impossible
to destroy the group later as we hung waiting for group's
->user_waits to reach 0 (which never happens due to the
leak.)  Fix the leak by calling fsnotify_finish_user_wait()
for all FANOTIFY_PERM_EVENTS that we have prepared for
user_wait.

Note that upstream has "inadvertently" fixed this in [1].
Cherry-picking it to older kernels requires some conflict
resolution and additional patches (but not too complicated),
except for v5.4 which is entirely too messy and complicated.

[1] commit 30ad1938326b ("fanotify: allow "weak" fsid when
    watching a single filesystem")
UPSTREAM-TASK=b:000000000
BUG=b:402438326
TEST=used a sample app from https://man7.org/linux/man-pages/man7/fanotify.7.html
     faked fanotify_get_fsid() fail and checked that group
     still could be destroyed
     under lockdep and kasan

Change-Id: I2fda4980cdfd6a3a47147e46e039b923360e4457
Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6589511
Reviewed-by: Tomasz Figa <tfiga@chromium.org>