CHROMIUM: iwl7000: mvm: avoid outdated reorder buffer head_sn
If no frames are received on a queue for a while, the reorder buffer
head_sn may be an old one. When the next frame that is received on
that queue and buffered is a subframe of an AMSDU but not the last
subframe, it will not update the buffer's head_sn. When the frame
release notification arrives, it will not release the buffered frame
because it will look like the notification's NSSN is lower than the
buffer's head_sn (because of a wraparound).
Fix it by updating the head_sn when the first frame is buffered.

BUG=b:360270690
TEST=WiFi_perf, wifi_matfunc

Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Change-Id: I9ec9a3508e7935e8d1833ea3e086066fdefee644
iwl7000-tree: 884e2f00a69d1ff6db4c619490800d7c98e237b1
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6746811
Tested-by: Miriam Rachel Korenblit <miriam.rachel.korenblit@intel.corp-partner.google.com>
Tested-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: David Ruth <druth@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Arowa Suliman <arowa@chromium.org>