Revert "CHROMIUM: drm/mediatek: clear payload id table before output the first stream"
Now, the DRM will automatically clear the payload ID table when plugging in the DP. There is no need to do anything extra in this case. If the scenario changes to a timing change, since we will allocate the payload again for both, it is not necessary to clear the payload.

However, we can still clear the payload before we re-allocate it. But using the code pair below will only send the sideband clear payload ID table without the DPCD 1c0 to 1c3 (0/0/0x3f), which will confuse the sink and cause some issues (for example, HDMI will always show black).
    mtk_dp->mgr.payload_id_table_cleared = false;
    drm_dp_mst_topology_queue_probe(&mtk_dp->mgr);

Hence, revert it first to fix those issues, and then create a new patch if we still want to clear the payload before we re-allocate it.

This reverts commit 50616e4e5f9d238c2252d93de83065253d666e5c.

BUG=b:421781781,b:421014104,b:427370441,b:427379747
TEST=build ok/verify pass locally
UPSTREAM-TASK=b:379039307

Signed-off-by: Ives Chenjh <ives.chenjh@mediatek.corp-partner.google.com>
Change-Id: Iac78769de58af880fd174fb7eea1d72dc1a5eca5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6712554
Reviewed-by: Jason-jh Lin <jason-jh.lin@mediatek.corp-partner.google.com>
Reviewed-by: Lancelot Wu <lancelot.wu@mediatek.corp-partner.google.com>
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>