CHROMIUM: iio: light: cros_ec_light_prox: Send ODR changes back to back
Due to a limitation of EC RGB sensor veml3328 driver, the ODR changes
must be sent back to back to the Clear and RGB sensors.
There is a window where older EC firmware could crash if the EC query
the sensor for RGB data while the ODR for the whole sensor (Clear + RGB) has already been set to 0.

BUG=b:424003439
UPSTREAM-TASK=b:294468604
TEST=On Xol, change sensor odr between 0Hz and 1Hz, observe there are no EC crash (version used: xol_14505.840.0)

Change-Id: I28af373acc4da5664a3884536170acfb05471ef1
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6715090
Reviewed-by: Boris Mittelberg <bmbm@google.com>