FROMGIT: HID: i2c-hid: Make elan touch controllers power on after panel is enabled
Introduce a new HID quirk to indicate that this device has to be enabled
after the panel's backlight is enabled, and update the driver data for
the elan devices to enable this quirk. This cannot be a I2C HID quirk
because the kernel needs to acknowledge this before powering up the
device and read the VID/PID. When this quirk is enabled, register
.panel_enabled()/.panel_disabling() instead for the panel follower.

Also rename the *panel_prepare* functions into *panel_follower* because
they could be called in other situations now.

Fixes: bd3cba00dcc63 ("HID: i2c-hid: elan: Add support for Elan eKTH6915 i2c-hid touchscreens")
Fixes: d06651bebf99e ("HID: i2c-hid: elan: Add elan-ekth6a12nay timing")
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Pin-yen Lin <treapking@chromium.org>
Acked-by: Jiri Kosina <jkosina@suse.com>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Link: https://lore.kernel.org/r/20250818115015.2909525-2-treapking@chromium.org
(cherry picked from commit cbdd16b818eef876dd2de9d503fe7397a0666cbe
 https://gitlab.freedesktop.org/drm/misc/kernel.git drm-misc-next)

BUG=b:305780363, b:424124093
TEST=emerge-skywalker sys-kernel/chromeos-kernel-6_6

Change-Id: I1b99cee6df254c4e8cdc3aea3875b0ef31a9cdb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6895537
Reviewed-by: Hsin-Te Yuan <yuanhsinte@chromium.org>
Commit-Queue: Pin-yen Lin <treapking@chromium.org>
Tested-by: Pin-yen Lin <treapking@chromium.org>