CHROMIUM: reven: Enable CONFIG_PINCTRL_METEORLAKE
This fixes track-pad on the HP EliteBook 660 G11.

The pins are not configured correctly. The I2C bus controller has no
physical path to communicate with the trackpad, so the trackpad is never
detected or fails to initialize.

With CONFIG_PINCTRL_METEORLAKE enabled The driver correctly configures
the pins, enabling the I2C bus and allowing the system to communicate
with the track-pad.

BUG=b:438758682
TEST=cros_workon start chromeos-kernel-6_6
TEST=emerge-reven chromeos-kernel-6_6
TEST=Boot on HP EliteBook 660 G11, verify track-pad works

Change-Id: I9c25edd021b1aba1e641467e95b5db8394b29d0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6880898
Commit-Queue: Jiao Zhou <jiaozhou@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Auto-Submit: Jiao Zhou <jiaozhou@google.com>
Tested-by: Jiao Zhou <jiaozhou@google.com>