CHROMIUM: drm/mediatek: Revising the RDMA FIFO value once more to address the intermittent black screen issue.
After conducting multiple experiments, it was discovered that when
setting the FIFO for RDMA, it needs to be divided by 16 and
left-shifted by 16 bits. Consequently, the previous setting of
16 * 2200 resulted in an oversized FIFO allocation for RDMA,
which did not adhere to the specifications, leading to a
potential black screen when switching resolutions.
Therefore, the FIFO size was adjusted to 2200, which
has been verified as a suitable value.

BUG=b:444086747
UPSTREAM-TASK=b:396344738
TEST=emerge-skywalker chromeos-kernel-6_6

Signed-off-by: Xiandong Wang <xiandong.wang@mediatek.corp-partner.google.com>
Change-Id: If783a00de6b5e64fb940016644a4793b591ea289
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6963234
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Reviewed-by: Hsin-Te Yuan <yuanhsinte@chromium.org>
Commit-Queue: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>