CHROMIUM: arm64: dts: mt8189-skywalker-anakin: Compatible with old keyboard
The old keyboard uses a 15*8 matrix, while the new keyboard uses
an 18*8 matrix. Since EC can only set one keyboard matrix
(can only be set to 18*8), in order to be compatible with
the old and new keyboards, the matrix of the old keyboard
is also changed to 18*8.

BUG=b:423412548
UPSTREAM-TASK=b:440262873

TEST=1. emerge-skywalker chromeos-kernel-6_6
     2. deploy kernel to DUT.
     3. verify keyboard

Change-Id: I26b6b67b33d538f9ccc6dc0d31905128693464ed
Signed-off-by: Zhaoxiong Lv <lvzhaoxiong@huaqin.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/6983542
Commit-Queue: Pin-yen Lin <treapking@chromium.org>
Reviewed-by: Tony Chang <tonytcc@google.com>
Reviewed-by: Wentao Qin <qinwentao@huaqin.corp-partner.google.com>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>